server/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
weasyprint>=62
Pillow>=10
python-dateutil>=2.9
markdown>=3.5
pyyaml>=6.0
